feat: add utils directory
This is where I'll keep little scripts and utility things for making
reltrans more interoperable with other languages and tools. Currently
this consists of a Makefile that is supposed to make building easier and
more portable, and a `cli.c` file for wrapping the resulting shared
library and providing a 'first-look' executable.